feat(customfunc/cidr): Add cidrHostEnd custom function
Add cidrHostEnd custom function to get addresses starting from the end
of the subnet and working backwards.

issue #37

This change adds a new custom function to the program which enables
users to get addresses similar to the existing cidrHost, except it
starts from the end of the addressible range and works backwards.

* Tags: cidrHostEnd cidr custom func subnet